separated que into primitive, fixed que bug across windows
