+++ src/ui/details/hidden_fields.rs
@@ -19,7 +19,6 @@ use crate::{
 			tooltip::TooltipSignals,
 		},
 	},
-	Que,
 };
 
 pub struct HiddeFields {
@@ -31,7 +30,6 @@ pub struct HiddeFields {
 	pub tooltip_signals: TooltipSignals,
 	pub set_list: WriteSignal<im::Vector<(usize, &'static str, usize)>>,
 	pub main_scroll_to: RwSignal<f32>,
-	pub que: Que,
 	pub env: Environment,
 }
 
@@ -45,7 +43,6 @@ pub fn hidden_fields(param: HiddeFields) -> impl View {
 		tooltip_signals,
 		set_list,
 		main_scroll_to,
-		que,
 		env,
 	} = param;
 	let is_expanded = create_rw_signal(false);
@@ -74,7 +71,6 @@ pub fn hidden_fields(param: HiddeFields) -> impl View {
 						is_hidden: true,
 						tooltip_signals,
 						set_list,
-						que,
 						env: env.clone(),
 					})
 					.style(|s| s.padding_bottom(5))
