+++ src/ui/details/list_item.rs
@@ -37,7 +37,6 @@ use crate::{
 			tooltip::TooltipSignals,
 		},
 	},
-	Que,
 };
 
 pub struct ListItem {
@@ -49,7 +48,6 @@ pub struct ListItem {
 	pub is_hidden: bool,
 	pub tooltip_signals: TooltipSignals,
 	pub set_list: WriteSignal<im::Vector<(usize, &'static str, usize)>>,
-	pub que: Que,
 	pub env: Environment,
 }
 
@@ -63,7 +61,6 @@ pub fn list_item(param: ListItem) -> impl View {
 		is_hidden,
 		tooltip_signals,
 		set_list,
-		que,
 		env,
 	} = param;
 
@@ -378,7 +375,6 @@ pub fn list_item(param: ListItem) -> impl View {
 			dates,
 			is_shown: !matches!(field, DbFields::Title),
 			field_title,
-			que,
 			tooltip_signals,
 			env: env_history,
 		}),
